Combo system V1
Added a simple combo system using switches to determine current attack and saving last attack with hard coded values for each combination. Hold Q to start a combo and release Q or let the enemy attack to shoot the combo. If a spell that isnt a part of the 4 elements is drawn while holding Q the combo will also shoot.